+++ ocfl/validator.py
@@ -18,7 +18,7 @@ from .constants import INVENTORY_FILENAME, SPEC_VERSIONS_SUPPORTED, \
 from .digest import file_digest, normalized_digest
 from .inventory_validator import InventoryValidator
 from .namaste import find_namastes
-from .pyfs import pyfs_openfs, pyfs_walk, pyfs_openfile, pyfs_files_identical
+from .fsw import fsw_openfs, fsw_walk, fsw_openfile, fsw_files_identical
 from .validation_logger import ValidationLogger
 
 
@@ -107,7 +107,7 @@ class Validator():
         return self.status_str()
 
     def validate_object(self, path):
-        """Validate OCFL object at path or pyfs root.
+        """Validate OCFL object at path or fsw root.
 
         Arguments:
             path: either a filepath or else an open fs filesystem
@@ -121,7 +121,7 @@ class Validator():
         self.initialize()
         try:
             if isinstance(path, str):
-                self.obj_fs = pyfs_openfs(path)
+                self.obj_fs = fsw_openfs(path)
             else:
                 self.obj_fs = path
                 path = self.obj_fs.to_json()  # FIXME - Better info?
@@ -130,7 +130,7 @@ class Validator():
             return False
         # Object declaration, set spec version number. If there are multiple declarations,
         # look for the lastest object version then report any others as errors
-        namastes = find_namastes(0, pyfs=self.obj_fs)
+        namastes = find_namastes(0, fsw=self.obj_fs)
         if len(namastes) == 0:
             self.log.error("E003a", assumed_version=self.spec_version)
         else:
@@ -146,7 +146,7 @@ class Validator():
                     self.log.error("E006", filename=namaste.filename)
                 elif spec_version is None or this_file_version > spec_version:
                     spec_version = this_file_version
-                    if not namaste.content_ok(pyfs=self.obj_fs):
+                    if not namaste.content_ok(fsw=self.obj_fs):
                         self.log.error("E007", filename=namaste.filename)
             if spec_version is None:
                 self.log.error("E003c", assumed_version=self.spec_version)
@@ -236,7 +236,7 @@ class Validator():
             digest_algorithm = m.group(1)
             try:
                 digest_recorded = self.read_inventory_digest(inv_digest_file)
-                digest_actual = file_digest(inv_file, digest_algorithm, pyfs=self.obj_fs)
+                digest_actual = file_digest(inv_file, digest_algorithm, fsw=self.obj_fs)
                 if digest_actual != digest_recorded:
                     self.log.error("E060", inv_file=inv_file, actual=digest_actual, recorded=digest_recorded, inv_digest_file=inv_digest_file)
             except ValueError as e:  # pylint: disable=broad-except
@@ -322,7 +322,7 @@ class Validator():
                 # Don't validate in this case. Per the spec the inventory in the last version
                 # MUST be identical to the copy in the object root, just check that
                 root_inv_file = INVENTORY_FILENAME
-                if not pyfs_files_identical(self.obj_fs, inv_file, root_inv_file):
+                if not fsw_files_identical(self.obj_fs, inv_file, root_inv_file):
                     self.log.error("E064", root_inv_file=root_inv_file, inv_file=inv_file)
                 else:
                     # We could also just compare digest files but this gives a more helpful error for
@@ -424,7 +424,7 @@ class Validator():
                         # Check content_directory
                         content_path = os.path.join(version_dir, self.content_directory)
                         num_content_files_in_version = 0
-                        for dirpath, dirs, files in pyfs_walk(self.obj_fs, content_path):
+                        for dirpath, dirs, files in fsw_walk(self.obj_fs, content_path):
                             if dirpath != "/" + content_path and (len(dirs) + len(files)) == 0:
                                 self.log.error("E024", where=version_dir, path=dirpath)
                             for file in files:
@@ -461,7 +461,7 @@ class Validator():
                         self.log.error("E092b", where="root", content_path=filepath)
                     else:
                         if self.check_digests:
-                            content_digest = file_digest(filepath, digest_type=self.digest_algorithm, pyfs=self.obj_fs)
+                            content_digest = file_digest(filepath, digest_type=self.digest_algorithm, fsw=self.obj_fs)
                             if content_digest != normalized_digest(digest, digest_type=self.digest_algorithm):
                                 self.log.error("E092a", where="root", digest_algorithm=self.digest_algorithm, digest=digest, content_path=filepath, content_digest=content_digest)
                             known_digests = {self.digest_algorithm: content_digest}
@@ -495,7 +495,7 @@ class Validator():
                     # Don't recompute anything, just use it if we've seen it before
                     content_digest = known_digests[digest_algorithm]
                 else:
-                    content_digest = file_digest(filepath, digest_type=digest_algorithm, pyfs=self.obj_fs)
+                    content_digest = file_digest(filepath, digest_type=digest_algorithm, fsw=self.obj_fs)
                     known_digests[digest_algorithm] = content_digest
                 for digest in additional_digests[filepath][digest_algorithm]:
                     if content_digest != normalized_digest(digest, digest_type=digest_algorithm):
@@ -512,12 +512,12 @@ class Validator():
             Exception: if there is an error reading the digest or it has
                 the wrong format
         """
-        with pyfs_openfile(inv_digest_file, "r", pyfs=self.obj_fs) as fh:
+        with fsw_openfile(inv_digest_file, "r", fsw=self.obj_fs) as fh:
             line = fh.readline()
             # we ignore any following lines, could raise exception
         m = re.match(r"""(\w+)\s+(\S+)\s*$""", line)
         if not m:
-            raise Exception("Bad inventory digest file %s, wrong format" % (inv_digest_file))
+            raise ValueError("Bad inventory digest file %s, wrong format" % (inv_digest_file))
         if m.group(2) != INVENTORY_FILENAME:
-            raise Exception("Bad inventory name in inventory digest file %s" % (inv_digest_file))
+            raise ValueError("Bad inventory name in inventory digest file %s" % (inv_digest_file))
         return m.group(1)