+++ tests/test_digest.py
@@ -44,10 +44,10 @@ class TestAll(unittest.TestCase):
         self.assertEqual(file_digest("tests/testdata/files/hello_out_there.txt", "md5"),
                          "9c7ec1389a61f1e15185bd976672bc63")
 
-    def test_file_digest__pyfs(self):
+    def test_file_digest__fsw(self):
         """Test file_digest method with content."""
         td_fs = DirFileSystem("tests/testdata", LocalFileSystem())
-        self.assertEqual(file_digest("files/hello_out_there.txt", "md5", pyfs=td_fs),
+        self.assertEqual(file_digest("files/hello_out_there.txt", "md5", fsw=td_fs),
                          "9c7ec1389a61f1e15185bd976672bc63")
 
     def test_string_digest(self):