+++ tests/test_object.py
@@ -5,7 +5,7 @@ import tempfile
 import unittest
 
 from ocfl.inventory import Inventory
-from ocfl.pyfs import pyfs_openfs, pyfs_listdir_names
+from ocfl.fsw import fsw_openfs, fsw_listdir_names
 from ocfl.object import Object, ObjectException
 from ocfl.version_metadata import VersionMetadata
 
@@ -59,7 +59,7 @@ class TestAll(unittest.TestCase):
     def test06_version_dirs_and_metadata(self):
         """Test version_dirs_and_metadata."""
         oo = Object(digest_algorithm="md5", spec_version='1.0')
-        src_fs = pyfs_openfs('fixtures/1.0/content/spec-ex-full')
+        src_fs = fsw_openfs('fixtures/1.0/content/spec-ex-full')
         vdirs = []
         metadatas = []
         for (vdir, metadata) in oo.version_dirs_and_metadata(src_fs):
@@ -75,17 +75,17 @@ class TestAll(unittest.TestCase):
 
     def test07_write_object_declaration(self):
         """Test write_object_declaration."""
-        tmpfs = pyfs_openfs("temp://")
+        tmpfs = fsw_openfs("temp://")
         oo = Object(obj_fs=tmpfs, spec_version='1.0')
         oo.write_object_declaration()
-        self.assertEqual(pyfs_listdir_names(tmpfs, '/'), ['0=ocfl_object_1.0'])
+        self.assertEqual(fsw_listdir_names(tmpfs, '/'), ['0=ocfl_object_1.0'])
 
     def test08_write_inventory_and_sidecar(self):
         """Test write_object_and_sidecar."""
-        tmpfs = pyfs_openfs("temp://")
+        tmpfs = fsw_openfs("temp://")
         oo = Object(obj_fs=tmpfs)
         oo.write_inventory_and_sidecar(Inventory({'abc': 'def'}))
-        self.assertEqual(set(pyfs_listdir_names(tmpfs,"/")),
+        self.assertEqual(set(fsw_listdir_names(tmpfs,"/")),
                          set(['inventory.json', 'inventory.json.sha512']))
         with tmpfs.open('inventory.json') as fh:
             j = json.load(fh)
@@ -96,7 +96,7 @@ class TestAll(unittest.TestCase):
         oo = Object(obj_fs=tmpfs)
         invdir = 'xxx'
         oo.write_inventory_and_sidecar(Inventory({'gh': 'ik'}), invdir)
-        self.assertEqual(set(pyfs_listdir_names(tmpfs, invdir)),
+        self.assertEqual(set(fsw_listdir_names(tmpfs, invdir)),
                          set(['inventory.json', 'inventory.json.sha512']))
         with tmpfs.open(os.path.join(invdir, 'inventory.json')) as fh:
             j = json.load(fh)